Task 5 second commit/added (float) to NumberDivider sout

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -9,7 +9,7 @@ public class NumberDivider {
 
         toBeDivided /=5;
 
-        System.out.println(toBeDivided);
+        System.out.println((float) toBeDivided);
 
     }
 }
